break out analyst modules, add charts feature
