--- lib/repository.rb
@@ -1,5 +1,0 @@
-module Repository
-  def method_name
-    a
-  end
-end
