+++ sites/charts.html
@@ -1,0 +1,68 @@
+<html>
+  <head>
+    <script type="text/javascript" src="https://www.gstatic.com/charts/loader.js"></script>
+    <script type="text/javascript">
+      google.charts.load("current", {packages:["corechart"]});
+      google.charts.setOnLoadCallback(drawMerchantRevenueChart);
+      google.charts.setOnLoadCallback(drawInvoiceDaysChart);
+      
+      function drawMerchantRevenueChart() {
+        var data = google.visualization.arrayToDataTable([
+          ['merchant_id', 'item_count'],
+          ["12334634", 0.21275227E6],
+          ["12334942", 0.20618195E6],
+          ["12334280", 0.20286397E6],
+          ["12336035", 0.19535152E6],
+          ["12335213", 0.17631668E6]
+        ]);
+
+        var options = {
+          title: 'Merchants With Highest Revenue',
+          pieHole: 0.4,
+        };
+
+        var chart = new google.visualization.PieChart(document.getElementById('donutchart'));
+        chart.draw(data, options);
+      }
+      
+      function drawInvoiceDaysChart() {
+      var data = google.visualization.arrayToDataTable([
+        ["Days", "Number of Invoices", { role: "style" } ],
+        ["Sunday", 708, "red"],
+        ["Monday", 696, "#orange"],
+        ["Tuesday", 692, "#yellow"],
+        ["Wednesday", 741, "#27ae60"]
+        ["Thursday", 718, "#1abc9c"],
+        ["Friday", 701, "#2980b9"],
+        ["Saturday", 729, "#8e44ad"]
+      ]);
+
+      var view = new google.visualization.DataView(data);
+      view.setColumns([0, 1,
+                       { calc: "stringify",
+                         sourceColumn: 1,
+                         type: "string",
+                         role: "annotation" },
+                       2]);
+
+      var options = {
+        title: "Number of Invoices by Day",
+        width: 600,
+        height: 400,
+        bar: {groupWidth: "95%"},
+        legend: { position: "none" },
+      };
+      var chart = new google.visualization.ColumnChart(document.getElementById("columnchart_values"));
+      chart.draw(view, options);
+  }
+    </script>
+  </head>
+  <body>
+    <table class = 'columns'
+      <tr>
+        <td><div id="donutchart" style="width: 900px; height: 500px;"></div></td>
+        <td><div id="columnchart_values" style="width: 900px; height: 300px;"></div></td>
+      </tr>
+    </table>
+  </body>
+</html>
